+++ htdocs/prestashopget/core/modules/modPrestashopGet.class.php
@@ -121,9 +121,7 @@ class modPrestashopGet extends DolibarrModules
 		// Example: $this->const=array(0=>array('PRESTASHOPGET_MYNEWCONST1','chaine','myvalue','This is a constant to add',1),
 		//                             1=>array('PRESTASHOPGET_MYNEWCONST2','chaine','myvalue','This is another constant to add',0, 'current', 1)
 		// );
-		$this->const = array(
-			1=>array('PRESTASHOPGET_MYCONSTANT', 'chaine', 'avalue', 'This is a constant to add', 1, 'allentities', 1)
-		);
+		$this->const = array();
 
 		// Some keys to add into the overwriting translation tables
 		/*$this->overwrite_translation = array(
@@ -209,25 +207,22 @@ class modPrestashopGet extends DolibarrModules
 
 		$r=0;
 		$this->rights[$r][0] = $this->numero + $r;	// Permission id (must not be already used)
-		$this->rights[$r][1] = 'Read myobject of PrestashopGet';	// Permission label
-		$this->rights[$r][3] = 1; 					// Permission by default for new user (0/1)
+		$this->rights[$r][1] = 'Read data from PrestashopGet';	// Permission label
 		$this->rights[$r][4] = 'read';				// In php code, permission will be checked by test if ($user->rights->prestashopget->level1->level2)
 		$this->rights[$r][5] = '';				    // In php code, permission will be checked by test if ($user->rights->prestashopget->level1->level2)
 
-		$r++;
+		/*$r++;
 		$this->rights[$r][0] = $this->numero + $r;	// Permission id (must not be already used)
 		$this->rights[$r][1] = 'Create/Update myobject of PrestashopGet';	// Permission label
-		$this->rights[$r][3] = 1; 					// Permission by default for new user (0/1)
 		$this->rights[$r][4] = 'write';				// In php code, permission will be checked by test if ($user->rights->prestashopget->level1->level2)
 		$this->rights[$r][5] = '';				    // In php code, permission will be checked by test if ($user->rights->prestashopget->level1->level2)
 
 		$r++;
 		$this->rights[$r][0] = $this->numero + $r;	// Permission id (must not be already used)
 		$this->rights[$r][1] = 'Delete myobject of PrestashopGet';	// Permission label
-		$this->rights[$r][3] = 1; 					// Permission by default for new user (0/1)
 		$this->rights[$r][4] = 'delete';				// In php code, permission will be checked by test if ($user->rights->prestashopget->level1->level2)
 		$this->rights[$r][5] = '';				    // In php code, permission will be checked by test if ($user->rights->prestashopget->level1->level2)
-
+		*/
 
 		// Main menu entries
 		$this->menu = array();			// List of menus to add
@@ -261,7 +256,7 @@ class modPrestashopGet extends DolibarrModules
 								'langs'=>'prestashopget@prestashopget',	        // Lang file to use (without .lang) by module. File must be in langs/code_CODE/ directory.
 								'position'=>1000+$r,
 								'enabled'=>'$conf->prestashopget->enabled',  // Define condition to show or hide menu entry. Use '$conf->prestashopget->enabled' if entry must be visible if module is enabled. Use '$leftmenu==\'system\'' to show if leftmenu system is selected.
-								'perms'=>'1',			                // Use 'perms'=>'$user->rights->prestashopget->level1->level2' if you want your menu with a permission rules
+								'perms'=>'$user->rights->prestashopget->read',			                // Use 'perms'=>'$user->rights->prestashopget->level1->level2' if you want your menu with a permission rules
 								'target'=>'',
 								'user'=>2);				                // 0=Menu for internal users, 1=external users, 2=both
 		/*$this->menu[$r++]=array(	'fk_menu'=>'fk_mainmenu=prestashopget,fk_leftmenu=prestashopget',	    // '' if this is a top menu. For left menu, use 'fk_mainmenu=xxx' or 'fk_mainmenu=xxx,fk_leftmenu=yyy' where xxx is mainmenucode and yyy is a leftmenucode